# apps/gateway/src/main/java/io/filternarrange/gateway/api/auth/dto/LoginRequest.java
// SPDX-License-Identifier: Apache-2.0
package io.filternarrange.gateway.api.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(@Email @NotBlank String email, @NotBlank String password) {}